server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pypdf2>=3.0.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.5.0

